requirements.txt
opencv-python
easyocr
torch
transformers
Pillow
numpy
sentencepiece
sacremoses
fastapi 
uvicorn[standard]
python-multipart